- Late fabs/fabsf() fixing.

+++ src/samplv1_lv2.cpp
@@ -181,7 +181,7 @@ void samplv1_lv2::run ( uint32_t nframes )
 						}
 						if (samplv1::paramValue(samplv1::DEL1_BPMSYNC) > 0.0f) {
 							const float bpm = samplv1::paramValue(samplv1::DEL1_BPM);
-							if (bpm > 0.0f && ::fabs(bpm_host - bpm) > 0.01f)
+							if (bpm > 0.0f && ::fabsf(bpm_host - bpm) > 0.01f)
 								samplv1::setParamValue(samplv1::DEL1_BPM, bpm_host);
 						}
 					}
